+++ tests/testthat/test-tdata.R
@@ -1,3 +1,5 @@
+withr::local_options(lifecycle_verbosity = "quiet")
+
 # ---- constructor ----
 testthat::test_that("new_tdata accepts reactive and not reactive MAE and data.frames", {
   utils::data(miniACC, package = "MultiAssayExperiment")
@@ -138,7 +140,7 @@ testthat::test_that("get_metadata returns NULL if dataset doesn't exist", {
 # ---- get_code ----
 testthat::test_that("get_code returns empty character if tdata object has no code", {
   my_tdata <- new_tdata(data = list(iris = iris, mtcars = mtcars))
-  testthat::expect_equal("", isolate(get_code(my_tdata)))
+  testthat::expect_equal("", isolate(get_code_tdata(my_tdata)))
 })
 
 testthat::test_that("get_code returns character of code if tdata object has code", {
@@ -149,14 +151,14 @@ testthat::test_that("get_code returns character of code if tdata object has code
     data = list(x = iris, mtcars = head(mtcars)),
     code = reactive(code_string)
   )
-  testthat::expect_equal(isolate(get_code(my_tdata)), code_string)
+  testthat::expect_equal(isolate(get_code_tdata(my_tdata)), code_string)
 
   # not reactive case (for constructor)
   my_tdata <- new_tdata(
     data = list(x = iris, mtcars = head(mtcars)),
     code = code_string
   )
-  testthat::expect_equal(isolate(get_code(my_tdata)), code_string)
+  testthat::expect_equal(isolate(get_code_tdata(my_tdata)), code_string)
 })
 
 # ---- get_code wrapper ----
@@ -199,7 +201,7 @@ testthat::test_that("tdata2env throws error if argument is not tdata", {
   testthat::expect_error(tdata2env(iris), "Must inherit from class 'tdata'")
 })
 
-# ---- get_join_keys ----
+# ---- join_keys ----
 testthat::test_that("join_keys returns NULL if no join_keys object exists inside tdata", {
   my_tdata <- new_tdata(data = list(iris = iris, mae = reactive(miniACC)))
   testthat::expect_null(join_keys(my_tdata))
@@ -217,4 +219,50 @@ testthat::test_that("join_keys returns join_keys object if it exists inside tdat
   )
 
   testthat::expect_equal(join_keys(my_tdata), jk)
+})
+
+
+# as_tdata ----
+code <- c("iris <- iris", "mtcars <- mtcars")
+data_tdata <- teal::new_tdata(list(iris = iris, mtcars = mtcars), code)
+data_teal_data <- teal.data::teal_data(iris = iris, mtcars = mtcars, code = code)
+data_reactive <- reactive(teal.data::teal_data(iris = iris, mtcars = mtcars, code = code))
+
+testthat::test_that("as_tdata accepts all possible inputs", {
+  testthat::expect_no_error(as_tdata(data_tdata))
+  testthat::expect_no_error(as_tdata(data_teal_data))
+  testthat::expect_no_error(as_tdata(data_reactive))
+})
+
+testthat::test_that("as_tdata always returns tdata object", {
+  data_tdata_downgraded <- as_tdata(data_tdata)
+  data_teal_data_downgraded <- as_tdata(data_teal_data)
+  data_reactive_downgraded <- as_tdata(data_teal_data)
+
+  testthat::expect_s3_class(data_tdata_downgraded, "tdata")
+  testthat::expect_s3_class(data_teal_data_downgraded, "tdata")
+  testthat::expect_s3_class(data_reactive_downgraded, "tdata")
+})
+
+testthat::test_that("datasets are maintained during conversion", {
+  data_tdata_downgraded <- as_tdata(data_teal_data)
+
+  datanames_teal_data <- sort(teal.data::datanames(data_teal_data))
+  datanames_tdata <- sort(names(data_tdata_downgraded))
+
+  testthat::expect_identical(datanames_teal_data, datanames_tdata)
+
+  datasets_teal_data <- sapply(datanames_teal_data, function(x) teal.code::get_var(data_teal_data, x))
+  datasets_tdata <- sapply(datanames_tdata, function(x) shiny::isolate(data_tdata_downgraded[[x]]()))
+
+  testthat::expect_identical(datasets_teal_data, datasets_tdata)
+})
+
+testthat::test_that("code is maintained during conversion", {
+  data_teal_data_downgraded <- as_tdata(data_teal_data)
+  skip("skipped until we resolve handling code in teal.data:::new_teal_data")
+  testthat::expect_identical(
+    teal.code::get_code(data_teal_data),
+    shiny::isolate(attr(data_teal_data_downgraded, "code")())
+  )
 })